Fixing broken prediction tests
The issue was due to now, BigML, with the new source server config,
says that in source `spam.csv` the field `Message` in `categorical`
instead of `text`. We have to update the source forcing this field as
`text`.
These tests are not completely fixed because, probably due to a bug in
BigML, the `term_analysis` settings of the field are lost. Once BigML
says what happens in this particular case, the tests will be fixed as
needed.